cleanup pytest_simcore
