+++ packages/settings-library/tests/test_email.py
@@ -9,6 +9,22 @@ import pytest
 from pydantic import ValidationError
 from settings_library.email import EmailProtocol, SMTPSettings
 
+from .helpers.typing_env import EnvVarsDict
+from .helpers.utils_envs import delenvs_from_dict
+
+
+@pytest.fixture
+def all_env_devel_undefined(
+    monkeypatch: pytest.MonkeyPatch, env_devel_dict: EnvVarsDict
+) -> None:
+    """Ensures that all env vars in .env-devel are undefined in the test environment
+
+    NOTE: this is useful to have a clean starting point and avoid
+    the environment to influence your test. I found this situation
+    when some script was accidentaly injecting the entire .env-devel in the environment
+    """
+    delenvs_from_dict(monkeypatch, env_devel_dict, raising=False)
+
 
 @pytest.mark.parametrize(
     "cfg",
